feat: add recurcion functional for 04 task

+++ 04-copy-directory/index.js
@@ -1,10 +1,10 @@
-const { mkdir, readdir, copyFile } = require('fs');
+const { mkdir, readdir, copyFile, stat } = require('fs');
 const path = require('path');
 
 const fromCopyDir = path.join(__dirname, '/files');
 
-function copyDir(pathFrom) {
-  const copyDirTo = pathFrom + '-copy';
+function copyDir(pathFrom, pathTo = '') {
+  const copyDirTo = pathTo ? pathTo : pathFrom + '-copy';
   mkdir(copyDirTo, { recursive: true }, (err) => {
     if (err) console.log(err);
 
@@ -12,13 +12,19 @@ function copyDir(pathFrom) {
       if (err) console.log(err);
 
       files.forEach((file) => {
-        copyFile(
-          path.join(pathFrom, file),
-          path.join(copyDirTo, file),
-          (error) => {
-            if (error) console.log(error);
-          },
-        );
+        stat(path.join(pathFrom, file), (error, stats) => {
+          if (stats.isFile()) {
+            copyFile(
+              path.join(pathFrom, file),
+              path.join(copyDirTo, file),
+              (error) => {
+                if (error) console.log(error);
+              },
+            );
+          } else {
+            copyDir(path.join(pathFrom, file), path.join(copyDirTo, file));
+          }
+        });
       });
     });
   });
